+++ src/VarDump/Visitor/Format/CollectionFormat.cs
@@ -1,0 +1,7 @@
+﻿namespace VarDump.Visitor.Format;
+
+public enum CollectionFormat
+{
+    MultiLine,
+    SingleLine
+}
