+++ test/VarDump.UnitTests/IntegralTypesSpec.cs
@@ -1,4 +1,5 @@
 using VarDump.Visitor;
+using VarDump.Visitor.Format;
 using Xunit;
 
 namespace VarDump.UnitTests;
